testing new root combine script

+++ run_Prod_Analysis.sh
@@ -539,14 +539,20 @@ if [[ $a_flag = "true" ]]; then
 	#python3 mergeRootFiles.py "${LTANAPATH}/OUTPUT/Analysis/${ANATYPE}LT/" "_-1_Raw_Data" "Cut_Kaon_Events_all_RF" "${OutDATAFilename}_Right" "${data_right[*]}"
 	#python3 mergeRootFiles.py "${LTANAPATH}/OUTPUT/Analysis/${ANATYPE}LT/" "_-1_Raw_Data" "Cut_Kaon_Events_prompt_RF" "${OutDATAFilename}_Right" "${data_right[*]}"
 	#python3 mergeRootFiles.py "${LTANAPATH}/OUTPUT/Analysis/${ANATYPE}LT/" "_-1_Raw_Data" "Cut_Kaon_Events_rand_RF" "${OutDATAFilename}_Right" "${data_right[*]}"
-	for i in "${data_right[@]}"
-	do
-	    cd "${LTANAPATH}/OUTPUT/Analysis/${ANATYPE}LT"
-	    echo "Combining run $i with ${OutDATAFilename}_Right.root..."
-	    echo "Renaming ${i}_Raw_Data to ${i}_Proc_Data..."
-	    #mv ${i}_-1_Raw_Data.root ${i}_-1_Proc_Data.root # <runNum>_-1_Proc_Data.root is used in later LT_analysis
-	    mv ${i}_-1_Proc_Data.root ${i}_-1_Raw_Data.root # <runNum>_-1_Proc_Data.root is used in later LT_analysis
-	done
+	if [ -f "${LTANAPATH}/OUTPUT/Analysis/${ANATYPE}LT/${OutDATAFilename}_Right.root" ]; then
+	    for i in "${data_right[@]}"
+	    do
+		if [ -f "${LTANAPATH}/OUTPUT/Analysis/${ANATYPE}LT/${i}_-1_Raw_Data.root" ]; then
+		    cd "${LTANAPATH}/OUTPUT/Analysis/${ANATYPE}LT"
+		    echo "Combining run $i with ${OutDATAFilename}_Right.root..."
+		    echo "Renaming ${i}_Raw_Data to ${i}_Proc_Data..."
+		    #mv ${i}_-1_Raw_Data.root ${i}_-1_Proc_Data.root # <runNum>_-1_Proc_Data.root is used in later LT_analysis
+		    mv ${i}_-1_Proc_Data.root ${i}_-1_Raw_Data.root # <runNum>_-1_Proc_Data.root is used in later LT_analysis
+		else
+		    echo "WARNING: ${i}_Raw_Data.root does not exist!"
+		fi
+	    done	 
+	fi
 	echo
     fi
